requirements.txt
uvicorn
fastapi
pymongo
motor
passlib
python-jose
pymongo
pydantic
fastapi-standalone-docs